refactor(chat): ChatInputBar 拆分命令解析 / 输入状态机 / resize
ChatInputBar.tsx 479 → ~240 行，退化为瘦渲染组件。

- utils/parse-command.ts：纯命令解析 parseChatCommand → 判别式结果
  （unknown / commandNoArgs / askNeedsQuestion / reviewAction / run / agentAsk），无 i18n / 副作用
- hooks/useChatInput.ts：输入栏状态机 hook —— 输入 / 提交（调 parseChatCommand 映射回调）/
  自动补全浮层 / 历史回放（shell 式 Up/Down）/ 停止请求 + 相关 effect
- hooks/useTextareaAutosizeDrag.ts：textarea 顶边拖拽调高
- ChatInputBar.tsx：消费两个 hook 做渲染（补全列表 / textarea / 命令栏 / 自动评审 / 发送停止）

纯结构拆分，行为不变；ChatInputBar 导出名与路径不变（ChatPane 引用零波及）。
